+++ packages/carbon-graphs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-*Changed 
+* Changed 
  * Fixed rendering of timeline graph when it is rendered below another graph.
  
 ## 2.16.3 - (March 16, 2021)