+++ packages/carbon-graphs/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
+
 * Changed 
   * Fixed rendering of timeline graph when it is rendered below another graph.
   * Fixed Y-axis and y2-axis icons not updating during reflow.
+  * Updated `packages/carbon-graphs/README.md` so that links point to updated documentation path.
 
 ## 2.16.3 - (March 16, 2021)